Generalize from temperature sensors
The information about the devices was previously published under the
metric `p3e_fb_temperator_sensor_info`: this is in general wrong as
there may be other devices attached to the FRITZ!Box as well. So rename
that metric to `p3e_fb_device_info`.

Also move the code for the general `p3e_fb_battery_status` metric right
below the code for `p3e_fb_device_info` as both are generic.